Separate out pvr_txr API.
Split from #539. Splits out the textures API.
Additionally removed some tabs that snuck in.
